Extract table view delegate and datasource from TimerListController Extract timer list view model delegate from TimerListController
