+++ ios/FarmTimer/FarmTimer/View/TimerList/TimerListController.swift
@@ -16,48 +16,12 @@ class TimerListController: UIViewController {
     override func viewDidLoad() {
         super.viewDidLoad()
         
+        setupTableViewDelegate()
+        setupTableViewDataSource()
+        bindTimerListViewModel()
+    }
+    
+    func setupUI() {
         tableView.tableFooterView = UIView()
-        
-        tableView.delegate = self
-        tableView.dataSource = self
-        listViewModel.delegate = self
     }
-    
-}
-
-extension TimerListController: UITableViewDataSource {
-    
-    func numberOfSections(in tableView: UITableView) -> Int {
-        return listViewModel.numberOfSections
-    }
-    
-    func tableView(_ tableView: UITableView, numberOfRowsInSection section: Int) -> Int {
-        return listViewModel.numberOfItems(at: section)
-    }
-    
-    func tableView(_ tableView: UITableView, cellForRowAt indexPath: IndexPath) -> UITableViewCell {
-        return tableView.dequeueReusableCell(withIdentifier: "TimerCell", for: indexPath)
-    }
-    
-    func tableView(_ tableView: UITableView, heightForRowAt indexPath: IndexPath) -> CGFloat {
-        return 80
-    }
-}
-
-extension TimerListController: UITableViewDelegate {
-    
-    func tableView(_ tableView: UITableView, willDisplay cell: UITableViewCell, forRowAt indexPath: IndexPath) {
-        guard let cell = cell as? TimerCell else { return }
-        guard let item = listViewModel.itemAt(row: indexPath.row, section: indexPath.section) else { return }
-        cell.titleLabel.text = item.title
-    }
-    
-}
-
-extension TimerListController: TimerListViewModelDelegate {
-    
-    func listViewAdded(_ model: TimerListViewModel) {
-        tableView.reloadData()
-    }
-    
 }
